Fix problem: JpaAuditing not working; Add & Refactor Exceptions
