+++ seminar2/src/main/kotlin/com/wafflestudio/seminar/SeminarAssignmentTwo.kt
@@ -2,7 +2,9 @@ package com.wafflestudio.seminar
 
 import org.springframework.boot.autoconfigure.SpringBootApplication
 import org.springframework.boot.runApplication
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing
 
+@EnableJpaAuditing
 @SpringBootApplication
 class SeminarOneApplication
