+++ seminar2/src/main/kotlin/com/wafflestudio/seminar/core/user/service/AuthException.kt
@@ -1,6 +1,21 @@
 package com.wafflestudio.seminar.core.user.service
 
-import com.wafflestudio.seminar.common.SeminarException
-import org.springframework.http.HttpStatus
+import com.wafflestudio.seminar.common.*
 
-class AuthException(val msg: String = "") : SeminarException(msg, HttpStatus.UNAUTHORIZED)
+typealias AuthException = Seminar401
+
+// 401
+object InvalidTokenException :
+    AuthException("Authorization token is in invalid form (unsupported, malformed, etc.)")
+object FailedToLogInException :
+    AuthException("Failed to log in")
+object AuthTokenExpiredException :
+    AuthException("Authorization token is expired")
+
+// 404
+object UserNotFoundException :
+    Seminar404("Cannot find user info")
+
+// 409
+object DuplicateEmailException :
+    Seminar409("Given email is already signed up")
